IOS-624-copy Make number of items function work

+++ src/CattyTests/Functions/List/NumberOfItemsFunctionTest.swift
@@ -57,7 +57,7 @@ class NumberOfElementsFunctionTest: XCTestCase {
         userVariableString.isList = true
         userVariableString.value = ["a", "b", "c"]
         
-        XCTAssertEqual(Double(4), function.value(parameter: userVariableString as AnyObject), accuracy: 0.0001)
+        XCTAssertEqual(Double(3), function.value(parameter: userVariableString as AnyObject), accuracy: 0.0001)
         
         // empty list
         let userVariableEmpty = UserVariable()
